+++ Changelog.md
@@ -5,6 +5,10 @@ Bug fixes
 
 * Use `__send__` rather than `send` to prevent naming collisions (Bradley Schaefer)
 
+Deprecations
+
+* Deprecates the `--webrat` option to the scaffold and request spec generator (Andy Lindeman)
+
 ### 2.99.0.beta1 / 2013-11-07
 [full changelog](http://github.com/rspec/rspec-rails/compare/v2.14.0...v2.99.0.beta1)